make handleCurse helper func and move gsit related events
